pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recon-filter"
version = "1.0.0"
description = "Professional Stream Processing Engine for System Logs & Recon Data"
readme = "README.md"
authors = [{name = "krvst"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
]
keywords = ["cli", "filtering", "security", "logs", "recon", "json", "csv"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "PyYAML>=6.0.0",
    "questionary>=2.0.0",
    "pypdf>=3.17.0",
    "charset-normalizer>=3.3.0",
    "psutil>=5.9.0",
    "ijson>=3.2.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
recon-filter = "recon_filter.main:app"

[tool.setuptools]
packages = ["recon_filter", "recon_filter.cli", "recon_filter.engine", "recon_filter.engine.handlers", "recon_filter.security"]
